Build wheel from sdist
Starting in version 0.5.0, the default behavior of `pypa/build` is to
first build the sdist, then to buil a wheel from the sdist, see:

    https://pypa-build.readthedocs.io/en/latest/changelog.html#id10

If you specify `-s -w`, both will be built from the local directory.
Since the sdist is the correct release artifact from which to build
wheels, the default behavior is preferred.